fix warning

+++ src/framework/graphics/image.h
@@ -38,7 +38,7 @@ public:
     void overwrite(const Color& color);
     void blit(const Point& dest, const ImagePtr& other);
     void paste(const ImagePtr& other);
-    void resize(const Size& size) { m_size = size; m_pixels.resize(size.area() * m_bpp, 0); }
+    void resize(const Size& size) { m_size = size; m_pixels.resize(size.area() * static_cast<size_t>(m_bpp), 0); }
     bool nextMipmap();
 
     void flipVertically();
@@ -54,7 +54,7 @@ public:
     int getWidth() { return m_size.width(); }
     int getHeight() { return m_size.height(); }
     int getBpp() { return m_bpp; }
-    uint8_t* getPixel(int x, int y) { return &m_pixels[(y * m_size.width() + x) * m_bpp]; }
+    uint8_t* getPixel(int x, int y) { return &m_pixels[static_cast<size_t>(y * m_size.width() + x) * m_bpp]; }
 
     bool hasTransparentPixel() const { return m_transparentPixel; }
     void setTransparentPixel(const bool value) { m_transparentPixel = value; }
